Edit - logonComponent
- Удален компонент login
- Компонент logon перерименован в login
