+++ src/app/views/hcms/hcms-layout/hcms-layout.component.html
@@ -5,7 +5,6 @@
       <nz-tab><a *nzTabLink nz-tab-link [routerLink]="['timesheet']">Расписание</a></nz-tab>
       <nz-tab><a *nzTabLink nz-tab-link [routerLink]="['analysis']">Аналитика</a></nz-tab>
     </nz-tabset>
-    <app-login class="ml-auto"></app-login>
   </nz-header>
   <nz-content class="p-3">
     <router-outlet></router-outlet>
