+++ src/app/views/hcms/hcms.module.ts
@@ -1,7 +1,6 @@
 import { NgModule }                 from '@angular/core';
 import { CommonModule }             from '@angular/common';
 import { ReactiveFormsModule }      from '@angular/forms';
-import { LoginModule }              from '@widgets/login/login.module';
 import { ReactiveInputModule }      from '@widgets/reactive/reactive-fields/reactive-input';
 import { ReactiveDatePickerModule } from '@widgets/reactive/reactive-fields/reactive-date-picker';
 import { ReactiveTextAreaModule }   from '@widgets/reactive/reactive-fields/reactive-text-area';
@@ -44,7 +43,6 @@ const ANT_DESIGN_MODULES = [
     CommonModule,
     ReactiveFormsModule,
     HcmsRoutingModule,
-    LoginModule,
     REACTIVE_FIELDS_MODULES,
     ANT_DESIGN_MODULES
   ],
