+++ src/app/views/welcome/views/login/login.component.html
@@ -1,11 +1,11 @@
-<ng-container *ngIf="logon$ | async"></ng-container>
+<ng-container *ngIf="login$ | async"></ng-container>
 
 <div class="d-flex justify-content-center w-100">
   <nz-card style="width: 400px" class="mt-3">
     <h4 class="text-center">Авторизуйтесь!</h4>
-    <app-reactive-input fieldId="logon"
+    <app-reactive-input fieldId="login"
                         label="Имя пользователя"
-                        [control]="fg.controls.logon">
+                        [control]="fg.controls.login">
     </app-reactive-input>
     <app-reactive-input fieldId="password"
                         label="Пароль"
@@ -13,8 +13,8 @@
                         [control]="fg.controls.password">
     </app-reactive-input>
     <button nz-button nzBlock
-            [nzLoading]="isLogonLoading"
-            (click)="onLogon()">Войти
+            [nzLoading]="isLoginLoading"
+            (click)="onLogin()">Войти
     </button>
   </nz-card>
 </div>
