+++ src/app/views/welcome/welcome-routing.module.ts
@@ -1,10 +1,10 @@
 import { NgModule }              from '@angular/core';
 import { RouterModule, Routes }  from '@angular/router';
-import { LogonComponent }        from './views/logon/logon.component';
+import { LoginComponent }        from './views/login/login.component';
 import { RegistrationComponent } from './views/registration/registration.component';
 
 const routes: Routes = [
-  {path: '', component: LogonComponent},
+  {path: '', component: LoginComponent},
   {path: 'registration', component: RegistrationComponent},
 ];
