+++ src/app/views/welcome/welcome.module.ts
@@ -1,11 +1,12 @@
-import { NgModule }              from '@angular/core';
-import { CommonModule }          from '@angular/common';
-import { ReactiveInputModule }   from '@widgets/reactive/reactive-fields/reactive-input';
-import { NzButtonModule }        from 'ng-zorro-antd/button';
-import { NzCardModule }          from 'ng-zorro-antd/card';
-import { RegistrationComponent } from './views/registration/registration.component';
-import { LogonComponent }        from './views/logon/logon.component';
-import { WelcomeRoutingModule }  from './welcome-routing.module';
+import { NgModule }                from '@angular/core';
+import { CommonModule }            from '@angular/common';
+import { ReactiveInputModule }     from '@widgets/reactive/reactive-fields/reactive-input';
+import { ValidationMessageModule } from '@widgets/validation-message';
+import { NzButtonModule }          from 'ng-zorro-antd/button';
+import { NzCardModule }            from 'ng-zorro-antd/card';
+import { RegistrationComponent }   from './views/registration/registration.component';
+import { LoginComponent }          from './views/login/login.component';
+import { WelcomeRoutingModule }    from './welcome-routing.module';
 
 const REACTIVE_FORMS_MODULES = [
   ReactiveInputModule
@@ -19,12 +20,13 @@ const ANT_DESIGN_MODULES = [
   imports: [
     CommonModule,
     WelcomeRoutingModule,
+    ValidationMessageModule,
     ANT_DESIGN_MODULES,
     REACTIVE_FORMS_MODULES
   ],
   declarations: [
     RegistrationComponent,
-    LogonComponent
+    LoginComponent
   ]
 })
 export class WelcomeModule {
